+++ lua/kubectl/client/rust/lib.rs
@@ -30,8 +30,10 @@ mod cmd;
 mod dao;
 mod describe;
 mod events;
+mod filter;
 mod log;
 mod processors;
+mod sort;
 mod store;
 mod structs;
 mod utils;
